added SpatialLocationField
