+++ mapbox_location_field/models.py
@@ -1,30 +1,10 @@
-from django.core.exceptions import ValidationError
+from django.contrib.gis.db.models import PointField
 from django.db import models
 from django.utils.translation import ugettext_lazy as _
 
-from .forms import AddressAutoHiddenField as AddressAutoHiddenFormField
+from .forms import AddressAutoHiddenField as AddressAutoHiddenFormField, parse_location
 from .forms import LocationField as LocationFormField
-
-
-def parse_location(location_string):
-    """parse and convert coordinates from string to tuple"""
-    args = location_string.split(",")
-    if len(args) != 2:
-        raise ValidationError(_("Invalid input for a Location instance"))
-
-    lat = args[0]
-    lng = args[1]
-
-    try:
-        lat = float(lat)
-    except ValueError:
-        raise ValidationError(_("Invalid input for a Location instance. Latitude must be convertible to float "))
-    try:
-        lng = float(lng)
-    except ValueError:
-        raise ValidationError(_("Invalid input for a Location instance. Longitude must be convertible to float "))
-
-    return lat, lng
+from .forms import SpatialLocationField as SpatialLocationFormField
 
 
 class LocationField(models.CharField):
@@ -82,3 +62,22 @@ class AddressAutoHiddenField(models.TextField):
         defaults = {'form_class': AddressAutoHiddenFormField}
         defaults.update(kwargs)
         return models.Field.formfield(self, **defaults)
+
+
+class SpatialLocationField(PointField):
+    description = _("Location field for spatial databases, stores Points.")
+
+    def __init__(self, *args, **kwargs):
+        self.map_attrs = kwargs.pop("map_attrs", {})
+        super().__init__(*args, **kwargs)
+
+    def deconstruct(self):
+        name, path, args, kwargs = super().deconstruct()
+        kwargs["map_attrs"] = self.map_attrs
+        return name, path, args, kwargs
+
+    def formfield(self, **kwargs):
+        defaults = {'form_class': SpatialLocationFormField}
+        defaults.update(kwargs)
+        defaults.update({"map_attrs": self.map_attrs})
+        return super().formfield(**defaults)
